database added and file restructed
