+++ .gitignore
@@ -1,3 +1,5 @@
 __pycache__/
 .pytest_cache/
 env/
+app/site.db
+app/__pycache__/
