+++ app/templates/login.html
@@ -1,0 +1,54 @@
+<!DOCTYPE html>
+{% extends "base.html" %}
+{% block content%}
+<div class="content-section">
+    <h1>Login!</h1>
+    <form action="" method="post">
+        {{ form.hidden_tag() }}
+        <fieldset class="form-group">
+            <legend>Login now</legend>
+            <div class="form-group">
+                {{ form.email.label(class="form-control-label")}}
+                {% if form.email.errors %}
+                {{form.email(class="form-control form-control-lg is-invalid")}}
+                <div class="invalid-feedback">
+                    {% for error in form.email.errors %}
+                        <span>{{ error }}</span>
+                    {% endfor %}
+                </div>
+            {% else %}
+                {{ form.email(class="form-control form-control-lg")}}
+            {% endif %}
+            </div>
+            <div class="form-group">
+                {{ form.password.label(class="form-control-label")}}
+                {% if form.password.errors %}
+                {{form.password(class="form-control form-control-lg is-invalid")}}
+                <div class="invalid-feedback">
+                    {% for error in form.password.errors %}
+                        <span>{{ error }}</span>
+                    {% endfor %}
+                </div>
+            {% else %}
+                {{ form.password(class="form-control form-control-lg")}}
+            {% endif %}
+            </div>
+            <div class="form-check">
+                {{ form.remember(class="form-check-input")}}
+                {{ form.remember.label(class="form-check-label") }}
+            </div>
+        </fieldset>
+        <div class="form-group">
+            {{ form.submit(class="btn btn-primary mt-2")}}
+        </div>
+        <small class="text-muted ml-2">
+            <a href="#">Forgot Password?</a>
+        </small>
+    </form>
+</div>
+<div class="border-top pt-3">
+    <small class="text-muted">
+        Need an account? <a class="ml-2" href="{{ url_for('register') }}">Sign up</a>
+    </small>
+</div>
+{% endblock %}
