+++ requirements.txt
@@ -2,3 +2,4 @@ pytest==6.1.2
 Flask==2.0.2
 Flask-WTF==1.0.0
 email_validator==1.1.3
+flask-sqlalchemy==2.5.1
