--- templates/login.html
@@ -1,29 +1,0 @@
-<!DOCTYPE html>
-{% extends "base.html" %}
-{% block content%}
-<div class="content-section">
-    <h1>Login!</h1>
-    <form action="" method="post">
-        {{ form.hidden_tag() }}
-        <fieldset class="form-group">
-            <legend>Login now</legend>
-            <div class="form-group">
-                {{ form.email.label(class="form-control-label")}}
-                {{ form.email(class="form-control form-control-lg")}}
-            </div>
-            <div class="form-group">
-                {{ form.password.label(class="form-control-label")}}
-                {{ form.password(class="form-control form-control-lg")}}
-            </div>
-        </fieldset>
-        <div class="form-group">
-            {{ form.submit(class="btn btn-primary mt-2")}}
-        </div>
-    </form>
-</div>
-<div class="border-top pt-3">
-    <small class="text-muted">
-        Need an account? <a class="ml-2" href="{{ url_for('register') }}">Sign Up Now</a>
-    </small>
-</div>
-{% endblock %}
