--- templates/register.html
@@ -1,38 +1,0 @@
-<!DOCTYPE html>
-{% extends "base.html" %}
-{% block content%}
-    <div class="content-section">
-        <h1>Register!</h1>
-        <form action="" method="post">
-            {{ form.hidden_tag() }}
-            <fieldset class="form-group">
-                <legend>Register now</legend>
-                <div class="form-group">
-                    {{ form.username.label(class="form-control-label")}}
-                    {{ form.username(class="form-control form-control-lg")}}
-                </div>
-                <div class="form-group">
-                    {{ form.email.label(class="form-control-label")}}
-                    {{ form.email(class="form-control form-control-lg")}}
-                </div>
-                <div class="form-group">
-                    {{ form.password.label(class="form-control-label")}}
-                    {{ form.password(class="form-control form-control-lg")}}
-                </div>
-                <div class="form-group">
-                    {{ form.confirm_password.label(class="form-control-label")}}
-                    {{ form.confirm_password(class="form-control form-control-lg")}}
-                </div>
-
-            </fieldset>
-            <div class="form-group">
-                {{ form.submit(class="btn btn-primary mt-2")}}
-            </div>
-        </form>
-    </div>
-    <div class="border-top pt-3">
-        <small class="text-muted">
-            Already have an account? <a class="ml-2" href="{{ url_for('login') }}">Sign In</a>
-        </small>
-    </div>
-{% endblock %}
